Front-End/Client - Stall on OpenCL
The front-end/client now waits for all commands issued against its boost::compute::command_queue to finish before considering the frame to be finished.

This includes the duration of everything running asynchronously in OpenCL in the frame duration measurement, making it more meaningful.

This can probably be removed once the values the pipeline generates are actually consumed: As indicated in the comment in the code the dynfu::pipeline_value objects representing those values will have to block until the values become available in OpenCL, thereby including the time for the kernels to complete in the measurement.