dynfu::kinect_fusion_opencl_measurement_pipeline_block - Tests Fix
The tests for the dynfu::kinect_fusion_opencl_measurement_pipeline_block class were failing with the new build system due to the fact the tests referred to the "cl" directory simply as "cl".  This worked under the old build system as under the old build system the working directory when running bin/tests (or bin/tests.exe) was the root of the repository.  However the new build system sets the working directory of the test executable to the "bin" directory when running it.  The "bin" directory does not contain a "cl" directory (let alone one containing the appropriate OpenCL source files) and therefore the files corresponding to the requested OpenCL program(s) could not be found.

Used dynfu::current_executable_parent_path to make the aforementioned tests immune to this form of error.