dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Make dynfu::opencl_vector_pipeline_value_extractor Objects Class Members
This prevents them from being created each time dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block::operator () is invoked.

If the extracted values live in a compatible boost::compute::vector on the GPU this shouldn't result in large savings.  If the extracted values live on the CPU or in an incompatible boost::compute::vector on the GPU this could result in large savings as memory doesn't have to be allocated each frame.