dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Comment on No Transpose
When the A matrix is retrieved from OpenCL it is not transposed, even though a transpose is required to convert between Eigen's matrix representations and those we use in OpenCL.  This is because the A matrix is guaranteed to be symmetric.  Added a comment to this effect so it would not be suspected of being a bug in the future.