dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Dynamic OpenCL Work Group Size in Tests
If the OpenCL implementation being used has a maximum work group size <16 then the maximum work group size is chosen, otherwise 16 (the default) is used.

Not doing this was causing issues when using the AMD CPU OpenCL implementation on Travis CI (which apparently doesn't support a work group size of 16).