dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - __global => __constant
More semantic and appears to yield a mild speed up (~83 ms for 15 iterations on a 640x480 frame to ~79 ms).

Also added a comment so that all arguments to the correspondences kernel are properly indicated with their argument number.

Tried to make the first argument to the parallel_sum kernel __constant but that caused OpenCL to have a heard attack.  I suspect this is because that pointer/buffer is used mutably elsewhere.  Admittedly I didn't put much time into verifying/investigating this.