dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Store Matrices in __local Memory
Splits the correspondences kernel into two phases:

1.	Load matrices from __global into __local memory
2.	Find correspondences

This appears to yield a variable speed up.  A group size of 16 appears to yield a ~2 ms speed up on my computer with 15 iterations on a 640x480 frame.  Larger group sizes appear to yield a larger performance benefit to a point (32 was another ~1 ms, 640 was a degradation).