dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Accept Matrices by Value in struct
Introduced the mat4 and mat3 structures to the pose_estimation.cl file.  These structures contain a 4x4 or 3x3 matrix (respectively) as an array of floats.  The correspondences kernel now accepts t_gk_prev_inverse, t_z, and k in this manner rather than by pointer to a global buffer.  This eliminates the need to manage global buffers and should reduce accesses to global memory.