dynfu::kinect_fusion_opencl_update_reconstruction_pipeline_block - OpenCL NaN Clean Up
TSDF values can no longer be NaN (the TSDF is initialized with 1s) and therefore some of the NaN checks make no sense.

The one check that does make sense is to check the value from the incoming depth frame (invalid measurements are indicated with NaN), which was moved to right after that load occurs rather than much later in the kernel.