dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block - Group Size 1 Fix
Due to the fact all integers are zero in modulo 1 dynfu::kinect_fusion_opencl_pose_estimation_pipeline_block would allocate two equally sized buffers for the output matrices when the group size was 1.  This was unnecessary due to the fact the parallel sum (which is the only thing that used both buffers at the same time) would never run in these circumstances.